refactor(pull out createManager, strip internals in tsdocs)
