+++ libs/react/src/index.ts
@@ -1,1 +1,2 @@
 export * from './lib/react';
+export { Manager } from './lib/manager';
