+++ package.json
@@ -28,8 +28,8 @@
     "deploy-website": "yarn api-docs && cd apps/rx-store-website && ../../node_modules/.bin/docusaurus deploy",
     "build-website": "yarn api-docs && cd apps/rx-store-website && ../../node_modules/.bin/docusaurus build",
     "api-docs": "yarn api-docs:core && yarn api-docs:react",
-    "api-docs:core": "rm -fr apps/rx-store-website/src/pages/api-core/ && typedoc --out apps/rx-store-website/src/pages/api-core/ --tsconfig libs/core/tsconfig.lib.json --theme default && rm -fr apps/rx-store-website/src/pages/api-core/assets",
-    "api-docs:react": "rm -fr apps/rx-store-website/src/pages/api-react/ && typedoc --out apps/rx-store-website/src/pages/api-react/ --tsconfig libs/react/tsconfig.lib.json --theme default  && rm -fr apps/rx-store-website/src/pages/api-react/assets"
+    "api-docs:core": "rm -fr apps/rx-store-website/src/pages/api-core/ && typedoc --stripInternal --out apps/rx-store-website/src/pages/api-core/ --tsconfig libs/core/tsconfig.lib.json --theme default && rm -fr apps/rx-store-website/src/pages/api-core/assets",
+    "api-docs:react": "rm -fr apps/rx-store-website/src/pages/api-react/ && typedoc --stripInternal --out apps/rx-store-website/src/pages/api-react/ --tsconfig libs/react/tsconfig.lib.json --theme default  && rm -fr apps/rx-store-website/src/pages/api-react/assets"
   },
   "dependencies": {
     "@angular/animations": "10.0.7",
